2003-06-24  Dalibor Topic <robilad@yahoo.com>
        * herrno.m4, in6addr.m4, include/dummyin6.h, replace/dummyin6.c,
        replace/getaddrinfo.c, replace/memset.c, sockaddrin6.m4,
        sockinttypes.m4: updated to getaddrinfo 1.6.

        Reported by:  Ito Kazumitsu <kaz@maczuka.gcd.org>

+++ include/dummyin6.h
@@ -64,17 +64,17 @@
 
 #ifndef HAVE_STRUCT_IN6_ADDR
 struct in6_addr {
-    unsigned char s6_addr[16];
+    unsigned char	s6_addr[16];
 };
 #endif
 
 #ifndef HAVE_STRUCT_SOCKADDR_IN6
 struct sockaddr_in6 {
-    sa_family_t sin6_family;
-    in_port_t sin6_port;
-    int sin6_flowinfo;
-    struct in6_addr sin6_addr;
-    int sin6_scope_id;
+    sa_family_t		sin6_family;
+    in_port_t		sin6_port;
+    unsigned long	sin6_flowinfo;
+    struct in6_addr	sin6_addr;
+    unsigned long	sin6_scope_id;
 };
 #endif
